+++ CHANGELOG.md
@@ -24,6 +24,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Seth Friedman](https://github.com/sethfri)
   [#11232](https://github.com/CocoaPods/CocoaPods/pull/11232)
 
+* Include subprojects in the plugin post-install hook context  
+  [Eric Amorde](https://github.com/amorde)
+  [#11224](https://github.com/CocoaPods/CocoaPods/pull/11224)
+
 ##### Bug Fixes
 
 * Clean sandbox when a pod switches from remote to local.